+++ R/refit.tidylda.R
@@ -9,13 +9,14 @@
 #' @param prior_weight Numeric, 0 or greater or \code{NA}. The weight of the 
 #'        \code{beta} as a prior from the base model. See Details, below.
 #' @param additional_k Integer number of topics to add, defaults to 0.
-#' @param optimize_alpha Logical. Do you want to optimize alpha every iteration?
-#'        Defaults to \code{FALSE}. See 'details' of documentation for
-#'        \code{\link[textmineR]{FitLdaModel}}for more information.
+#' @param additional_eta_sum Numeric magnitude of prior for additional topics.
+#'        Ignored if \code{additional_k} is 0. Defaults to 250.
+#' @param optimize_alpha Logical. Experimental. Do you want to optimize alpha
+#'        every iteration? Defaults to \code{FALSE}.
 #' @param calc_likelihood Logical. Do you want to calculate the log likelihood every iteration?
 #'        Useful for assessing convergence. Defaults to \code{FALSE}.
 #' @param calc_r2 Logical. Do you want to calculate R-squared after the model is trained?
-#'        Defaults to \code{FALSE}. This calls \code{\link[textmineR]{CalcTopicModelR2}}.
+#'        Defaults to \code{FALSE}.
 #' @param return_data Logical. Do you want \code{new_data} returned as part of the model object?
 #' @param threads Number of parallel threads, defaults to 1.
 #' @param verbose Logical. Do you want to print a progress bar out to the console?
@@ -65,8 +66,9 @@
 #'   to an integer greater than zero. New entries to \code{alpha} have a flat
 #'   prior equal to the median value of \code{alpha} in the old model. (Note that
 #'   if \code{alpha} itself is a flat prior, i.e. scalar, then the new topics have
-#'   the same value for their prior.) New entries to \code{eta} are the average
-#'   of all previous topics in \code{eta}.
+#'   the same value for their prior.) New entries to \code{eta} have a shape 
+#'   from the average of all previous topics in \code{eta} and scaled by
+#'   \code{additional_eta_sum}.
 #' @note
 #'  Updates are, as of this writing, are almost-surely useful but their behaviors
 #'  have not been optimized or well-studied. Caveat emptor!
@@ -74,7 +76,7 @@
 #' @examples
 #' \donttest{
 #' # load a document term matrix
-#' data(nih_sample_dtm, package = "textmineR")
+#' data(nih_sample_dtm)
 #'
 #' d1 <- nih_sample_dtm[1:50, ]
 #'
@@ -120,6 +122,7 @@ refit.tidylda <- function(
   burnin = -1,
   prior_weight = 1,
   additional_k = 0, 
+  additional_eta_sum = 250,
   optimize_alpha = FALSE, 
   calc_likelihood = FALSE,
   calc_r2 = FALSE, 
@@ -214,11 +217,14 @@ refit.tidylda <- function(
     Nv = ncol(object$beta)
   )  
   
+  
   # if necessary, re-scale so that new eta has the weight prescribed by prior-weight
   if (! is.na(prior_weight)) {
     w_star <- rowSums(object$counts$Cv) + rowSums(eta$eta)
     
     eta$eta <- prior_weight * w_star * object$beta
+    
+    eta$eta_class <- "matrix" # always a matrix for refits using eta as prior
   }
   
   dimnames(eta$eta) <- dimnames(object$beta)
@@ -283,6 +289,8 @@ refit.tidylda <- function(
   )
   
   m_add <- t(t(m_add) + colMeans(eta$eta))
+  
+  m_add <- m_add / rowSums(m_add) * additional_eta_sum
   
   eta$eta <- rbind(eta$eta, m_add) # add new topics to eta