+++ R/utils.R
@@ -448,7 +448,7 @@ initialize_topic_counts <- function(
 #'     was trained on normalized so that it sums to 100.
 #'   \code{coherence} makes a call to \code{\link[tidylda]{calc_prob_coherence}}
 #'     using the default 5 most-probable terms in each topic.
-#'   \code{top_terms} displays the top 3 most-probable terms in each topic.
+#'   \code{top_terms} displays the top 5 most-probable terms in each topic.
 #' @note
 #'   \code{prevalence} should be proportional to P(topic). It is calculated by
 #'   weighting on document length. So, topics prevalent in longer documents get
@@ -467,7 +467,11 @@ initialize_topic_counts <- function(
 summarize_topics <- function(theta, beta, dtm) {
   
   # probabilistic coherence with default value for m
-  coherence <- calc_prob_coherence(beta = beta, data = dtm)
+  if (nrow(dtm) == 1) {
+    coherence <- rep(NA, nrow(beta))
+  } else {
+    coherence <- calc_prob_coherence(beta = beta, data = dtm)
+  }
   
   # prevalence of each topic, weighted by terms
   prevalence <- Matrix::rowSums(dtm) * theta
@@ -478,7 +482,7 @@ summarize_topics <- function(theta, beta, dtm) {
   
   # top 3 terms
   top_terms <- apply(beta, 1, function(x) {
-    names(x)[order(x, decreasing = TRUE)][1:3]
+    names(x)[order(x, decreasing = TRUE)][1:5]
   })
   
   top_terms <- apply(top_terms, 2, function(x) {
@@ -890,7 +894,7 @@ calc_lambda <- function(beta, theta, p_docs = NULL, correct = TRUE){
 #'   All 6 differences are averaged together.
 #' @examples
 #' # Load a pre-formatted dtm and topic model
-#' data(nih_sample_dtm, package = "textmineR")
+#' data(nih_sample_dtm)
 #' 
 #' # fit a model
 #' set.seed(12345)