+++ src/lda_gibbs2.cpp
@@ -10,9 +10,16 @@
 // Export this as a header for use in other packages
 // [[Rcpp::interfaces(r, cpp)]] 
 
-#include "parallel_gibbs_utils.h"
-#include "sample_int.h"
+#define ARMA_64BIT_WORD 1
+#include <RcppArmadillo.h>
+// [[Rcpp::depends(RcppArmadillo)]]
+// [[Rcpp::plugins(cpp11)]] 
+
+#include "sample_int.h" 
 #include "matrix_conversions.h"
+#include "parallel_gibbs_utils.h"
+
+
 
 #include <RcppArmadillo.h>
 #define ARMA_64BIT_WORD
@@ -21,6 +28,8 @@
 
 #include <progress.hpp>
 #include <progress_bar.hpp>
+
+
 
 //' Make a lexicon for looping over in the gibbs sampler
 //' @keywords internal
@@ -130,7 +139,7 @@ Rcpp::List create_lexicon(
     
     Rcpp::checkUserInterrupt();    
   }
-
+  
   
   // ***************************************************************************
   // Calculate Cd, Cv, and Ck from the sampled topics
@@ -208,6 +217,7 @@ Rcpp::List create_lexicon(
     _["Ck"]   = Ck
   );
 }
+
 
 ////////////////////////////////////////////////////////////////////////////////
 // 
@@ -322,7 +332,7 @@ Rcpp::List fit_lda_c(
   for (auto k = 0; k < Nk; k++) {
     sum_eta[k] = std::accumulate(eta[k].begin(), eta[k].end(), 0.0);
   }
-
+  
   // For aggregating samples post burn in
   std::vector<std::vector<long>> Cd_sum(Nd);
   std::vector<std::vector<double>> Cd_mean(Nd);
@@ -333,7 +343,7 @@ Rcpp::List fit_lda_c(
       Cd_mean[d].push_back(0.0);
     }
   }
-
+  
   std::vector<std::vector<long>> Cv_sum(Nk);
   std::vector<std::vector<double>> Cv_mean(Nk);
   
@@ -479,7 +489,7 @@ Rcpp::List fit_lda_c(
       } // end loop over documents
     } // end loop over threads
     
-
+    
     // update global Ck and Cv using batch versions
     if (threads > 1) {
       Ck = update_global_Ck(
@@ -518,7 +528,7 @@ Rcpp::List fit_lda_c(
         }
       }
       
-
+      
       
       if (! freeze_topics) {
         for (auto k = 0; k < Nk; k++) {