+++ tests/testthat/test-tidy-methods.R
@@ -1,6 +1,6 @@
 context("tests of tidy methods for tidylda")
 
-dtm <- textmineR::nih_sample_dtm
+dtm <- nih_sample_dtm
 
 d1 <- dtm[1:50, ]
 
@@ -184,7 +184,7 @@ test_that("glance works with updated models", {
 # note this uses unigrams and bigrams to ensure that there isn't 
 # 100% overlap in vocabulary between data and model
 tidy_docs <- 
-  textmineR::nih_sample[1:10, ] %>% 
+  nih_sample[1:10, ] %>% 
   dplyr::select(APPLICATION_ID, ABSTRACT_TEXT) %>% 
   tidytext::unnest_tokens(output = word, 
                           input = ABSTRACT_TEXT,
@@ -195,14 +195,14 @@ tidy_docs <-
 test_that("augment.tidylda behaves nicely", {
   
   td <- tidy_docs
-  
-  colnames(td)[1:2] <- c("document", "term")
-  
+
   # tidy tibble input
   a <- augment(
     x = lda, 
     data = td,
-    type = "class"
+    type = "class",
+    document_col = colnames(td)[1],
+    term_col = colnames(td)[2]
   )
   
   # right number of rows?
@@ -211,7 +211,7 @@ test_that("augment.tidylda behaves nicely", {
   # correctly identified topics?
   expect_equal(
     a %>% 
-      dplyr::filter(term %in% colnames(lda$beta)) %>%
+      dplyr::filter(.[[colnames(td)[2]]] %in% colnames(lda$beta)) %>%
       dplyr::select(topic) %>%
       dplyr::summarise(na_topic = sum(is.na(topic))) %>%
       as.numeric,